SetCollisionChainDebugThreshold接口参数修改为int类型
**复现步骤**

无

**原因和解决方案**

- 需要和interface一致

**影响范围**

无

**验证方案**

单元测试通过

**涉及分支**

* master

**检查项**

- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [x] 编译通过
- [x] 单元测试通过

/assign @jinjie

/cc @yuanchao @xiangyang

+++ hmap/lru/u128_lru.go
@@ -289,8 +289,8 @@ func (m *U128LRU) GetCollisionChain() []byte {
 	return chain
 }
 
-func (m *U128LRU) SetCollisionChainDebugThreshold(t uint32) {
-	atomic.StoreUint32(&m.collisionChainDebugThreshold, t)
+func (m *U128LRU) SetCollisionChainDebugThreshold(t int) {
+	atomic.StoreUint32(&m.collisionChainDebugThreshold, uint32(t))
 }
 
 func (m *U128LRU) Get(key0, key1 uint64, peek bool) (interface{}, bool) {
